pyproject.toml
[tool.poetry]
name = "robotic-researcher"
version = "0.1.0"
description = "wikipedia bot - Quandri's technical assignment"
authors = ["Lucas Rocha <8rocha.lucas@gmail.com>"]
readme = "README.md"
packages = [{include = "robotic_researcher"}]

[tool.poetry.dependencies]
python = "^3.10"
rpaframework = "^23.0.0"
dateparser = "^1.1.8"
urllib3 = "<2"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
black = "^23.3.0"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
pythonpath = "."
markers = [
    "integration: marks tests as integration test (deselect with '-m \"not integration\"')",
]

[tool.black]
line-length = 120

[tool.isort]
atomic = true
profile = "black"

[build-system]
requires = ["setuptools", "poetry-core"]
build-backend = "poetry.core.masonry.api"
